Change: CargoApi translated to Spanish + Text Corrections in EmpleadoApi & VentaApi

+++ src/main/java/com/tiendadbii/tiendadbii/controller/CargoApi.java
@@ -1,14 +1,20 @@
 package com.tiendadbii.tiendadbii.controller;
 
 import com.tiendadbii.tiendadbii.dto.CargoDto;
+import com.tiendadbii.tiendadbii.dto.ErrorDto;
 import com.tiendadbii.tiendadbii.model.entity.Cargo;
 import com.tiendadbii.tiendadbii.model.service.interfaces.ICargoService;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.modelmapper.ModelMapper;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -18,25 +24,41 @@ import java.util.List;
 @RequiredArgsConstructor
 @CrossOrigin
 @RequestMapping("/api/cargo")
-@Tag(name = "Cargo", description = "API CRUD operations for Cargo")
+@Tag(name = "6. Cargo", description = "Con este EndPoint se puede realizar todas las operaciones de Cargo.")
 @Log4j2
 public class CargoApi {
   private final ICargoService cargoService;
   private final ModelMapper modelMapper;
 
-  @Operation(summary = "List all registered Cargo", description = "The search is performed in the database without filter")
+  @Operation(summary = "Listar todos los cargos", description = "Retorna la lista de todos los cargos sin aplicar filtro.")
   @GetMapping
   public ResponseEntity<List<CargoDto>> findAll() {
     return ResponseEntity.ok(cargoService.findAll().stream().map(this::toDto).toList());
   }
 
-  @Operation(summary = "Find Cargo with given ID", description = "Given an idCargo, it will return Cargo from DB")
+  @Operation(summary = "Buscar un Cargo dado {idCargo}", description = "Retorna un Cargo con todos sus campos.",
+    responses = {
+      @ApiResponse(responseCode = "200", description = "Cargo encontrado exitosamente"),
+      @ApiResponse(responseCode = "404", description = "Cargo no encontrado con el id: {idCargo}", content = @Content(mediaType = MediaType.TEXT_PLAIN_VALUE))
+    }
+  )
   @GetMapping("/{idCargo}")
   public ResponseEntity<CargoDto> findById(@PathVariable Integer idCargo) {
-    return ResponseEntity.ok(this.toDto(cargoService.findById(idCargo)));
+    Cargo cargo = cargoService.findById(idCargo);
+    if (cargo == null) {
+      throw new EntityNotFoundException("Cargo no encontrado con el id: " + idCargo);
+    }
+    return ResponseEntity.ok(this.toDto(cargo));
   }
 
-  @Operation(summary = "Create new Cargo", description = "Given an Cargo Dto, it will be created in the database, id must be null")
+  @Operation(summary = "Crear nuevo cargo", description = "Para crear un cargo debe proporcionar un valor a todos los campos a excepcion de " +
+    "<b>\"fechaRegistro\"</b> e <b>\"idCargo\"</b>, las cuales se les asignara un valor internamente de manera automática.",
+    responses = {
+      @ApiResponse(responseCode = "201", description = "Cargo creado exitosamente"),
+      @ApiResponse(responseCode = "400", description = "Error en la Request, asegurese de completar todos los campos de manera correcta", content = @Content(schema = @Schema(implementation = ErrorDto.class))),
+      @ApiResponse(responseCode = "500", description = "Error interno del servidor, mas detalles en campo \"message\"", content = @Content(schema = @Schema(implementation = ErrorDto.class)))
+    }
+  )
   @PostMapping
   public ResponseEntity<CargoDto> createCargo(@RequestBody CargoDto dto) {
     CargoDto cargoDto = this.toDto(cargoService.createNew(this.toEntity(dto)));
@@ -46,15 +68,36 @@ public class CargoApi {
       .body(cargoDto);
   }
 
-  @Operation(summary = "Update Cargo", description = "Given a Cargo Dto, the Cargo will be updated in the database, id must not be null")
+  @Operation(summary = "Actualizar cargo", description = "Todos los valores de un cargo pueden ser actualizados, a excepción del campo " +
+    "<b>\"fechaRegistro\"</b>",
+    responses = {
+      @ApiResponse(responseCode = "200", description = "Cargo actualizado exitosamente"),
+      @ApiResponse(responseCode = "404", description = "Cargo no encontrado con el id: {idCargo}", content = @Content(mediaType = MediaType.TEXT_PLAIN_VALUE, schema = @Schema(implementation = String.class))),
+      @ApiResponse(responseCode = "500", description = "Error interno del servidor, mas detalles en campo \"message\"", content = @Content(schema = @Schema(implementation = ErrorDto.class)))
+    }
+  )
   @PutMapping
   public ResponseEntity<CargoDto> updateCargo(@RequestBody CargoDto dto) {
+    Cargo cargo = cargoService.findById(dto.getIdCargo());
+    if (cargo == null) {
+      throw new EntityNotFoundException("Cargo no encontrado con el id: " + dto.getIdCargo());
+    }
     return ResponseEntity.ok(this.toDto(cargoService.update(this.toEntity(dto))));
   }
 
-  @Operation(summary = "Delete Cargo", description = "Given a Cargo Dto, the Cargo will be deleted and the response will be empty, id must not be null")
+  @Operation(summary = "Eliminar Cargo", description = "Eliminara el registro del cargo, sin embargo, para realizar la operación, el cargo no debe estar asociado a ningún empleado",
+    responses = {
+      @ApiResponse(responseCode = "204", description = "Cargo eliminado exitosamente"),
+      @ApiResponse(responseCode = "404", description = "Cargo no encontrado con el id: {idCargo}", content = @Content(mediaType = MediaType.TEXT_PLAIN_VALUE, schema = @Schema(implementation = String.class))),
+      @ApiResponse(responseCode = "500", description = "Error interno del servidor, mas detalles en campo \"message\"", content = @Content(schema = @Schema(implementation = ErrorDto.class)))
+    }
+  )
   @DeleteMapping("/{idCargo}")
-  public ResponseEntity<CargoDto> deleteCargo(@PathVariable Integer idCargo) {
+  public ResponseEntity<Void> deleteCargo(@PathVariable Integer idCargo) {
+    Cargo cargo = cargoService.findById(idCargo);
+    if (cargo == null) {
+      throw new EntityNotFoundException("Cargo no encontrado con el id: " + idCargo);
+    }
     cargoService.deleteById(idCargo);
     return ResponseEntity.noContent().build();
   }
